a2a_search/requirements.txt
pydantic-ai 
mcp 
langchain-mcp-adapters

pydantic-graph
arxiv
google-generativeai
python-dotenv
requests
rich
asyncclick
flask
a2a-sdk
langchain 
langgraph
google-genai 
httpx 
langchain-google-genai 
uvicorn 
click 
rich
